requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
xgboost>=2.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0
Pillow>=10.2.0
scikit-image>=0.22.0
tensorflow>=2.15.0
